# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Docuflow API
  version: 0.2.0
  description: Local-first Go API. Original PDFs are immutable and processing outputs are versions.
servers:
  - url: /api
paths:
...
  /documents:
    get:
      operationId: listDocuments
      responses:
        "200":
          description: Document dashboard data
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Immutable original and upload version created
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document: { $ref: "#/components/schemas/Document" }
        default: { $ref: "#/components/responses/Error" }
...
  /documents/trash:
    get:
      operationId: listTrash
      description: Lists soft-deleted documents whose immutable bytes are still retained.
      responses:
        "200":
          description: Recoverable Trash records
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
        default: { $ref: "#/components/responses/Error" }
  /documents/bulk-delete:
    post:
      operationId: bulkDeleteDocuments
      description: >-
        Moves several documents to Trash in one call. Originals and versions stay
        on disk. The response is itemised: a document that is already gone is
        reported as a failure without affecting the rest of the selection.
      requestBody: { $ref: "#/components/requestBodies/BulkDocuments" }
      responses:
        "200": { $ref: "#/components/responses/BulkDeleteResult" }
        default: { $ref: "#/components/responses/Error" }
  /documents/bulk-permanent-delete:
    post:
      operationId: bulkPermanentlyDeleteDocuments
      description: >-
        Permanently removes several documents already in Trash, including their
        original and every version. Each document is staged and committed on its
        own, so the ones reported as deleted are fully gone and the ones reported
        as failed keep both their rows and their bytes. This cannot be undone.
      requestBody: { $ref: "#/components/requestBodies/BulkDocuments" }
      responses:
        "200": { $ref: "#/components/responses/BulkDeleteResult" }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: getDocument
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document: { $ref: "#/components/schemas/Document" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: renameDocument
      description: Changes application display metadata only; the UUID-backed storage path is unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string, maxLength: 240, example: "proposal-final.pdf" } }
      responses:
        "200": { description: Display name updated }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteDocument
      description: Soft deletes document metadata from Recent Files. Immutable original and version bytes are retained.
      responses:
        "204": { description: Document removed from Recent Files }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/restore:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    post:
      operationId: restoreDocument
      description: Restores a soft-deleted document to Recent Files.
      responses:
        "204": { description: Document restored }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/permanent:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    delete:
      operationId: permanentlyDeleteDocument
      description: Permanently removes a Trash record, its original PDF, and every stored version. Active documents cannot use this endpoint.
      responses:
        "204": { description: Document and stored bytes permanently removed }
        default: { $ref: "#/components/responses/Error" }
...
  /documents/{documentId}/versions:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: listDocumentVersions
      responses:
        "200":
          description: Newest-first immutable version history
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions: { type: array, items: { $ref: "#/components/schemas/DocumentVersion" } }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/metadata:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: getDocumentMetadata
      responses:
        "200": { description: Editable PDF metadata and read-only document information }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateDocumentMetadata
      parameters:
        - name: X-Confirm-Signature-Invalidation
          in: header
          schema: { type: string, enum: ["true", "false"] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PDFMetadata" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /documents/{documentId}/pages/{page}/thumbnail:
    parameters:
      - { $ref: "#/components/parameters/DocumentId" }
      - { name: page, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      operationId: getPageThumbnail
      responses:
        "200": { description: Lazily rendered PNG thumbnail, content: { image/png: { schema: { type: string, format: binary } } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  requestBodies:
    BulkDocuments:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [documentIds]
            properties:
              documentIds:
                type: array
                minItems: 1
                maxItems: 200
                items: { type: string, format: uuid }
                description: Repeated identifiers are collapsed to a single removal.
...
  responses:
    Error:
      description: Stable error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ProcessingResult:
      description: Output committed as a new immutable version
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DirectToolResult" }
...
    BulkDeleteResult:
      description: Per-document outcome for one bulk deletion
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BulkDeleteResult" }
  schemas:
    BulkDeleteResult:
      type: object
      required: [deleted, failed]
      properties:
        deleted:
          type: array
          items: { type: string, format: uuid }
          description: Documents that were removed. Empty rather than null.
        failed:
          type: array
          description: Documents left untouched, each with the reason. Empty rather than null.
          items:
            type: object
            required: [documentId, code, message]
            properties:
              documentId: { type: string, format: uuid }
              code: { type: string, example: DOCUMENT_NOT_FOUND }
              message: { type: string }
...
    DirectToolResult:
      type: object
      required: [document, version, downloadUrl, beforeBytes, afterBytes, savedToRecent]
      properties:
        document: { $ref: "#/components/schemas/Document" }
        version: { $ref: "#/components/schemas/DocumentVersion" }
        downloadUrl: { type: string, example: "/api/documents/00000000-0000-0000-0000-000000000000/versions/00000000-0000-0000-0000-000000000000/content" }
        beforeBytes: { type: integer, format: int64 }
        afterBytes: { type: integer, format: int64 }
        savedToRecent: { type: boolean, description: The current local persistence implementation records direct uploads internally. }
        outputName: { type: string, description: Suggested download filename for an individual output. }
...
    Document:
      type: object
      required: [id, originalName, mediaType, byteSize, checksumSha256, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        originalName: { type: string }
        mediaType: { type: string, const: application/pdf }
        byteSize: { type: integer, format: int64 }
        pageCount: { type: [integer, "null"], minimum: 1 }
        checksumSha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, "null"], format: date-time }
    DocumentVersion:
      type: object
      required: [id, documentId, parentVersionId, operation, byteSize, checksumSha256, metadata, createdAt]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        parentVersionId: { type: [string, "null"], format: uuid }
        operation: { type: string }
        byteSize: { type: integer, format: int64 }
        checksumSha256: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
...
    PDFMetadata:
      type: object
      required: [title, author, subject, keywords]
      properties:
        title: { type: string }
        author: { type: string }
        subject: { type: string }
        keywords: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }